+++ source_gen/source_gen/example/example.g.dart
@@ -1,5 +1,5 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
-// 2015-02-16T01:28:06.435Z
+// 2015-02-16T18:15:06.117Z
 
 part of source_gen.example.example;
 
@@ -49,3 +49,32 @@ Item _$ItemFromJson(Map<String, Object> json) => new Item()
   ..count = json['count']
   ..itemNumber = json['itemNumber']
   ..isRushed = json['isRushed'];
+
+// **************************************************************************
+// Generator: Instance of 'JsonLiteralGenerator'
+// Target: glossaryData
+// **************************************************************************
+
+final _$glossaryDataJsonLiteral = {
+  "glossary": {
+    "title": "example glossary",
+    "GlossDiv": {
+      "title": "S",
+      "GlossList": {
+        "GlossEntry": {
+          "ID": "SGML",
+          "SortAs": "SGML",
+          "GlossTerm": "Standard Generalized Markup Language",
+          "Acronym": "SGML",
+          "Abbrev": "ISO 8879:1986",
+          "GlossDef": {
+            "para":
+                "A meta-markup language, used to create markup languages such as DocBook.",
+            "GlossSeeAlso": ["GML", "XML"]
+          },
+          "GlossSee": "markup"
+        }
+      }
+    }
+  }
+};