+++ source_gen/source_gen/lib/generators/json_literal.dart
@@ -1,0 +1,8 @@
+library source_gen.example.json_literal_annotation;
+
+// TODO: option to make the literal a const
+class JsonLiteral {
+  final String path;
+
+  const JsonLiteral(this.path);
+}